Move sub-classes of "GridErrorDlg", "HotCommands" and "Label" classes to their own files
